Stop suffixing build Python with .exe on macOS
This is an ancient Python hack that is intended to deconflict the Python
binary produced in the source tree from the `Python` source directory on
a case-insensitive filesystem.  Unfortunately, `meson` treats all `.exe`
files on non-Windows systems as .NET assemblies and tries to run them
with `mono` which is obviously wrong and makes it impossible to build
meson-based Python packages.